+++ rastervision/data/raster_source/rasterio_source.py
@@ -10,14 +10,18 @@ from rastervision.data.raster_source import RasterSource
 from rastervision.core.box import Box
 
 
-def load_window(image_dataset, window=None, channels=None, is_masked=False):
-    """Load a window of an image from a TIFF file.
+def load_window(image_dataset, window=None, is_masked=False):
+    """Load a window of an image using Rasterio.
 
     Args:
+        image_dataset: a Rasterio dataset
         window: ((row_start, row_stop), (col_start, col_stop)) or
         ((y_min, y_max), (x_min, x_max))
-        channels: An optional list of bands to read.
-        is_masked: If True, read a  masked array from rasterio
+        is_masked: If True, read a masked array from rasterio
+
+    Returns:
+        np.ndarray of shape (height, width, channels) where channels is the number of
+            channels in the image_dataset.
     """
     if is_masked:
         im = image_dataset.read(window=window, boundless=True, masked=True)
@@ -30,14 +34,20 @@ def load_window(image_dataset, window=None, channels=None, is_masked=False):
         if nodata is not None and nodata != 0:
             im[channel, im[channel] == nodata] = 0
 
-    if channels:
-        im = im[channels, :]
     im = np.transpose(im, axes=[1, 2, 0])
     return im
 
 
 class RasterioRasterSource(ActivateMixin, RasterSource):
     def __init__(self, raster_transformers, temp_dir, channel_order=None):
+        """Constructor.
+
+        If channel_order is None, then use non-alpha channels. This also sets any
+        masked or NODATA pixel values to be zeros.
+
+        Args:
+            channel_order: list of indices of channels to extract from raw imagery
+        """
         self.temp_dir = temp_dir
         self.image_temp_dir = None
         self.image_dataset = None
@@ -45,11 +55,17 @@ class RasterioRasterSource(ActivateMixin, RasterSource):
 
         # Activate in order to get information out of the raster
         with self.activate():
-            colorinterp = self.image_dataset.colorinterp
-            self.channels = [
-                i for i, color_interp in enumerate(colorinterp)
-                if color_interp != ColorInterp.alpha
-            ]
+            num_channels = self.image_dataset.count
+            if channel_order is None:
+                colorinterp = self.image_dataset.colorinterp
+                if colorinterp:
+                    channel_order = [
+                        i for i, color_interp in enumerate(colorinterp)
+                        if color_interp != ColorInterp.alpha
+                    ]
+                else:
+                    channel_order = list(range(0, num_channels))
+            self.validate_channel_order(channel_order, num_channels)
 
             mask_flags = self.image_dataset.mask_flag_enums
             self.is_masked = any(
@@ -57,19 +73,12 @@ class RasterioRasterSource(ActivateMixin, RasterSource):
 
             self.height = self.image_dataset.height
             self.width = self.image_dataset.width
-            # Get 1x1 chip (after applying raster transformers) to test dtype
-            # and channel order if needed
+
+            # Get 1x1 chip and apply raster transformers to test dtype.
             test_chip = self.get_raw_chip(Box.make_square(0, 0, 1))
-
-            raw_channels = list(range(0, test_chip.shape[2]))
-            self.channel_order = channel_order or raw_channels
-            num_channels = len(raw_channels)
-
-            # Transform the chip to get the final dtype
-            test_chip = test_chip[:, :, self.channel_order]
+            test_chip = test_chip[:, :, channel_order]
             for transformer in raster_transformers:
                 test_chip = transformer.transform(test_chip, channel_order)
-
             self.dtype = test_chip.dtype
 
             self._set_crs_transformer()
@@ -95,8 +104,10 @@ class RasterioRasterSource(ActivateMixin, RasterSource):
     def _get_chip(self, window):
         if self.image_dataset is None:
             raise ActivationError('RasterSource must be activated before use')
-        return load_window(self.image_dataset, window.rasterio_format(),
-                           self.channels)
+        return load_window(
+            self.image_dataset,
+            window=window.rasterio_format(),
+            is_masked=self.is_masked)
 
     def _activate(self):
         # Download images to temporary directory and delete it when done.